Use .bashrc instead of .bash_profile, bash/ -> shell/, add zsh config
Using `.bashrc` as the source of all bash config
instead of `.bash_profile`, `.bash_profile` simply sources
`.bashrc` now. This make sure that the config gets
loaded even in non-login shells, such as when started
with the plain `$ bash` command.
This might not be needed for bash now, but now
when experimenting with zsh, and mostly just starting
it from a bash session, we need to have it's config
in `.zshrc`. So this makes sure that the bash config
is symmetrical to that.