Fix bash prompt leaving some characters to it
Similar fix was done in e966548.
Reference: https://unix.stackexchange.com/a/28828/337515

Also, moving the command to the beginning of the prompt so that it gets
executed first and thus the title gets reset a tiny bit faster.

Note that there's no need to add \[ \] to all prompt functions, it's
only needed in those that output non-printable characters.