Fix IdeaVim `go` not working with Python generic classes
For example with:
```
class FooQuerySet(QuerySet["Foo"]):
```
`go` would go to start of `Foo` when it should go to the start of
`QuerySet`.

Would ideally use something like this:
```
nnoremap go ?\(^class .\{-}([a-z.]\+\)\@<=[A-Z]<CR>:action GotoDeclarationOnly<CR>:noh<CR>
```
but unfortunately IdeaVim bugs out and leaves the first `<CR>` and
everything after it as just part of the search query.