Fix $PATH not being set correctly in MacVim
This was broken all the time, but only manifested on a fresh macOS
installation which didn't have a random leftover `npm` symlink in
`/usr/local/bin`.

This also makes it so that finally $FZF_IGNORES work correctly in
MacVim, and we don't get tons of bad search suggestions.

Fix from:
https://superuser.com/q/31353/999318
https://github.com/macvim-dev/macvim/blob/e4c45796accc63db497852d086b923b98ee14dd3/runtime/doc/gui_mac.txt#L297